No more time limit for stored email addresses
It has become clear that it is essential for the product support team to know which add-ons belong to which users. This cannot be determined effectively without storing each user's email address as a human-readable (and verifiable) personal identifier. Therefore, the privacy policy no longer puts a time limit on the storage of Heroku add-on user email addresses.